1.添加相关配置文件以及解析配置参数 2.修改命令行参数 3.修正并发抄表bail直接返回的问题

+++ src/service/plc_init.rs
@@ -32,16 +32,16 @@ impl PlcInit {
     pub fn new(
         uart_msg_sender: mpsc::Sender<UartMessage>,
         services: ModuleService,
-        timeout: u64,
-        init_timeout: u64,
+        timeout: u32,
+        init_timeout: u32,
     ) -> Self {
         Self {
             uart_msg_sender,
             services,
             init_result: Mutex::new(PlcInitResult::default()),
             cond: Condvar::new(),
-            timeout: Duration::from_millis(timeout),
-            init_timeout: Duration::from_secs(init_timeout),
+            timeout: Duration::from_millis(timeout as u64),
+            init_timeout: Duration::from_secs(init_timeout as u64),
             init_flag: Arc::new(AtomicBool::new(false)),
         }
     }
